+++ CHANGELOG.md
@@ -9,8 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Support ES Modules for config, migrations, adapter and plugins via `--es-modules` flag
-- Support for transpiled languages via `--sourceDir`, `--sourceMigrationExtension`,
-`--migrationExtension` (should also work with `ts-node`)
+- Support for transpiled languages via `--source-dir`, `--source-migration-extension`,
+`--migration-extension` (should also work with `ts-node`)
 - TypeScript type declarations for all exposed entities (`Adapter` interface
 and `MigrationManager` class)
 - Builtin TypeScript migration template file